Add a testbed for chaining SRv6-unaware VNfs
We provide a simple VirtualBox testbed using Vagrant. The testbed is
composed of three Virtual Machines (VMs) that represent SR ingress node,
NFV node, and SR egress node. VNFs are running as network namespaces in the
NFV node. For simplicity, The SR proxy behavior uses the same interface for
Target and source interface (possible only in the case of unidirectional SFC)